Use hmis-api, create base controller
