+++ config/routes.rb
@@ -655,21 +655,6 @@ Rails.application.routes.draw do
     end
   end
 
-  if ENV['ENABLE_HMIS_API'] == 'true'
-    namespace :hmis, defaults: { format: :json } do
-      devise_for :users, class_name: 'HmisUser',
-                         skip: [:registrations, :invitations, :passwords, :confirmations, :unlocks],
-                         path: '', path_names: { sign_in: 'login', sign_out: 'logout' }
-
-      resources :user, only: [:none] do
-        get :index, on: :collection
-      end
-      resources :projects, only: [:none] do
-        post :index, on: :collection
-      end
-    end
-  end
-
   resources :hmis, only: [:index, :show]
 
   resources :weather, only: [:index]
@@ -900,5 +885,21 @@ Rails.application.routes.draw do
     get :details
     get :actioncable
   end
+
+  if ENV['ENABLE_HMIS_API'] == 'true'
+    namespace :hmis_api, path: 'hmis-api', defaults: { format: :json } do
+      devise_for :users, class_name: 'HmisApiUser',
+                         skip: [:registrations, :invitations, :passwords, :confirmations, :unlocks, :password_expired],
+                         path: '', path_names: { sign_in: 'login', sign_out: 'logout' }
+
+      resources :user, only: [:none] do
+        get :index, on: :collection
+      end
+      resources :projects, only: [:none] do
+        post :index, on: :collection
+      end
+    end
+  end
+
   root 'root#index'
 end
